cleanup of old combatant code
