+++ _src/store.js
@@ -1,7 +1,7 @@
 import Vue from "vue";
 import Vuex from "vuex";
 import campaign from "@/models/campaign";
-import combatant from "@/models/combatant_mc";
+import combatant from "@/models/combatant";
 import creatureprop from "@/models/creatureprop";
 import encounter from "@/models/encounter";
 import gmscreentab from "@/models/gmscreentab";
